Moved logging functionality to a separate module
